--- server/api/baseModels.py
@@ -1,14 +1,0 @@
-from django.db import models
-
-
-
-
-class SoftDeleteModel(models.Model):
-    deleted = models.BooleanField(default=False)
-
-    def delete(self):
-        self.deleted = True
-        self.save()
-
-    class Meta:
-        abstract = True
